requirements.txt
# Core Dependencies
cryptography>=41.0.0
Pillow>=10.0.0
numpy>=1.24.0
requests>=2.31.0
flask>=3.0.0

# Optional for enhanced features
scipy>=1.11.0
